+++ timedf_benchmarks/hm_fashion_recs/preprocess.py
@@ -11,7 +11,7 @@ from typing import Any
 import logging
 
 from timedf import tm
-from timedf.pandas_backend import pd
+from timedf.backend import pd
 
 from . import schema
